api/requirements.txt
flask==2.3.3
elasticsearch==8.13.0
psycopg2-binary==2.9.9
faker==20.1.0